+++ src/main/java/com/flapkap/vendingmachine/web/request/ProductUpdateRequest.java
@@ -21,7 +21,7 @@ public record ProductUpdateRequest(
 
         @Schema(description = "Product cost", example = "1500.0")
         @PositiveOrZero(message = "error.product.invalidCost")
-        Double cost,
+        Integer cost,
 
         @Schema(description = "Product amount", example = "200")
         @PositiveOrZero(message = "error.product.invalidAmount")
